Optimize style layout
