+++ paimon-web-ui/package.json
@@ -12,13 +12,15 @@
   "dependencies": {
     "@douyinfe/semi-icons": "^2.38.1",
     "@douyinfe/semi-ui": "^2.38.1",
+    "@monaco-editor/react": "^4.5.1",
+    "@semi-bot/semi-theme-doucreator": "^1.0.18",
+    "@semi-bot/semi-theme-figma": "^0.1.6",
     "less": "^4.1.3",
-    "monaco-editor": "^0.39.0",
+    "monaco-editor": "^0.40.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.10.1",
-    "@monaco-editor/react": "^4.5.1"
+    "react-icons": "^4.10.1"
   },
   "devDependencies": {
     "@types/node": "^20.3.2",
