+++ paimon-web-ui/src/app/App.less
@@ -27,9 +27,4 @@ body,
   --semi-color-primary: rgba(var(--semi-grey-6), 1);
   --semi-color-primary-hover: rgba(var(--semi-violet-3), 1);
   --semi-color-primary-light-default: rgba(var(--semi-color-bg-0), 1);
-  .flex-between {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
 }
