+++ paimon-web-ui/src/pages/Layout/layout.module.less
@@ -15,16 +15,6 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
 
-import PlaygroundRightHeaderTab from "@pages/Playground/RightContent/TopPanel/Tab";
-
-const TopPanel = () => {
-    return(
-        <div>
-            <span>
-                <PlaygroundRightHeaderTab/>
-            </span>
-        </div>
-    )
+.content {
+    background-color: var(--semi-color-bg-0);
 }
-
-export default TopPanel;
