+++ paimon-web-ui/src/pages/Metadata/components/LeftContent/left-content.module.less
@@ -15,24 +15,17 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
 
-import { Layout } from "@douyinfe/semi-ui";
-import Header from "@pages/Layout/Header";
-import { Outlet } from 'react-router-dom';
-import styles from "./layout.module.less";
-
-const { Content } = Layout
-
-const LayoutRoot = () => {
-    return (
-        <Layout className="layout-page">
-            <Header/>
-            <Layout>
-                {/*<Sider/>*/}
-                <Content className={styles.content}>
-                    <Outlet />
-                </Content>
-            </Layout>
-        </Layout>
-    )
+.container {
+    width: 347px;
+    height: 100%;
+    padding: 10px 24px;
 }
-export default LayoutRoot
+
+.add-catalog-container {
+    color: var(--semi-color-text-0);
+    margin: 6px 0 16px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+}
