+++ paimon-web-ui/src/pages/Metadata/components/RightContent/right-content.module.less
@@ -15,20 +15,7 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
 
-import SiderTab from "@pages/Playground/TabMenuSidebar/SiderTab";
-
-const TabMenuSidebar = () => {
-    return(
-        <div style={{
-            position: 'fixed',
-            marginTop: '-23px',
-            marginLeft: '-25px',
-            backgroundColor: 'var(--semi-color-bg-1)',
-            width: '414px',
-            height: '100%'}}>
-            <SiderTab/>
-        </div>
-    )
+.container {
+  flex: 1;
+  padding: 10px 15px 0 15px;
 }
-
-export default TabMenuSidebar;
