+++ paimon-web-ui/src/pages/Playground/components/RightContent/components/BottomPanel/index.tsx
