+++ paimon-web-ui/src/pages/Playground/components/RightContent/right-content.module.less
@@ -14,3 +14,9 @@ software distributed under the License is distributed on an
 KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
+
+.container {
+  width: calc(100% - 416px);
+  background-color: var(--semi-color-bg-0);
+  padding: 0 0 0 2px;
+}
