+++ paimon-web-ui/src/pages/Playground/components/TabMenuSidebar/components/SiderTab/siderbar.module.less
@@ -15,24 +15,12 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
 
-import { Layout } from "@douyinfe/semi-ui";
-import Header from "@pages/Layout/Header";
-import { Outlet } from 'react-router-dom';
-import styles from "./layout.module.less";
-
-const { Content } = Layout
-
-const LayoutRoot = () => {
-    return (
-        <Layout className="layout-page">
-            <Header/>
-            <Layout>
-                {/*<Sider/>*/}
-                <Content className={styles.content}>
-                    <Outlet />
-                </Content>
-            </Layout>
-        </Layout>
-    )
+.container {
+  width: 100%;
+  margin-right: 5px;
 }
-export default LayoutRoot
+
+.tab-panel-container {
+  width: 100%;
+  padding: 10px 0 0 0;
+}
