+++ paimon-web-ui/src/pages/Playground/components/TabMenuSidebar/components/TreeComponent/index.tsx
@@ -15,51 +15,55 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
 
-import {Input, Tree} from '@douyinfe/semi-ui';
-import { IconSearch } from "@douyinfe/semi-icons";
+import { Tree } from '@douyinfe/semi-ui';
+import { Input } from '@douyinfe/semi-ui';
+import { IconFilter } from '@douyinfe/semi-icons';
+import { IconFile } from '@douyinfe/semi-icons';
+import styles from "./tree-component.module.less";
 
-const CatalogTree = () => {
+const TreeNode = () => {
     const treeData = [
         {
-            label: 'paimon_catalog_01',
-            value: 'catalog01',
+            label: 'paimon',
+            value: 'paimon',
             key: '0',
             children: [
                 {
-                    label: 'paimon_db_01',
-                    value: 'db01',
+                    label: 'paimon_table_01',
+                    value: 'paimon_table_01',
                     key: '0-0',
+                    icon: <IconFile/>,
                     children: [
                         {
-                            label: 'paimon_table_01',
-                            value: 'paimon_table_01',
+                            label: 'id',
+                            value: 'id',
                             key: '0-0-0',
                         },
                         {
-                            label: 'paimon_table_02',
-                            value: 'paimon_table_02',
+                            label: 'name',
+                            value: 'name',
                             key: '0-1-0',
                         },
                         {
-                            label: 'paimon_table_03',
-                            value: 'paimon_table_03',
+                            label: 'age',
+                            value: 'age',
                             key: '0-2-0',
                         },
                         {
-                            label: 'paimon_table_04',
-                            value: 'paimon_table_04',
+                            label: 'gender',
+                            value: 'gender',
                             key: '0-3-0',
                         },
                         {
-                            label: 'paimon_table_05',
-                            value: 'paimon_table_05',
+                            label: 'address',
+                            value: 'address',
                             key: '0-4-0',
                         },
                     ],
                 },
                 {
-                    label: 'paimon_db_02',
-                    value: 'paimon_db_02',
+                    label: 'paimon_table_02',
+                    value: 'paimon_table_02',
                     key: '0-1',
                     children: [
                         {
@@ -70,8 +74,8 @@ const CatalogTree = () => {
                     ]
                 },
                 {
-                    label: 'paimon_db_03',
-                    value: 'paimon_db_03',
+                    label: 'paimon_table_03',
+                    value: 'paimon_table_03',
                     key: '0-2',
                     children: [
                         {
@@ -82,8 +86,8 @@ const CatalogTree = () => {
                     ],
                 },
                 {
-                    label: 'paimon_db_04',
-                    value: 'paimon_db_04',
+                    label: 'paimon_table_04',
+                    value: 'paimon_table_04',
                     key: '0-3',
                     children: [
                         {
@@ -94,8 +98,8 @@ const CatalogTree = () => {
                     ],
                 },
                 {
-                    label: 'paimon_db_05',
-                    value: 'paimon_db_05',
+                    label: 'paimon_table_05',
+                    value: 'paimon_table_05',
                     key: '0-4',
                     children: [
                         {
@@ -108,8 +112,8 @@ const CatalogTree = () => {
             ],
         },
         {
-            label: 'paimon_catalog_02',
-            value: 'paimon_catalog_02',
+            label: 'iceberg',
+            value: 'iceberg',
             key: '1',
             children: [
                 {
@@ -126,10 +130,15 @@ const CatalogTree = () => {
         }
     ];
 
-    const style = {
-        width: '18vw',
-        height: '100%',
-    };
+    /**
+     * 渲染树节点 使用自定义方式 ,可以直接点击节点 展开
+     * @param className
+     * @param onExpand
+     * @param onClick
+     * @param data
+     * @param expandIcon
+     * @returns {JSX.Element}
+     */
 
     const renderLabel = (x: any) => {
         const className = x.className;
@@ -151,19 +160,18 @@ const CatalogTree = () => {
         );
     };
 
-
     return(
         <Tree
             filterTreeNode
             treeData={treeData}
             searchPlaceholder={"Filter"}
             searchRender={({ prefix, ...restProps }) => (
-                <Input suffix={<IconSearch />} {...restProps} style={{ width: '15vw', marginLeft: '-12px'}}></Input>
+                <Input suffix={<IconFilter/>} {...restProps}></Input>
             )}
             renderFullLabel={renderLabel}
-            style={style}
+            searchClassName={styles['tree-search-wrapper']}
         />
     )
 }
 
-export default CatalogTree;
+export default TreeNode;
