+++ paimon-web-ui/src/pages/Playground/playground.module.less
@@ -15,24 +15,13 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License. */
 
-import { Layout } from "@douyinfe/semi-ui";
-import Header from "@pages/Layout/Header";
-import { Outlet } from 'react-router-dom';
-import styles from "./layout.module.less";
-
-const { Content } = Layout
-
-const LayoutRoot = () => {
-    return (
-        <Layout className="layout-page">
-            <Header/>
-            <Layout>
-                {/*<Sider/>*/}
-                <Content className={styles.content}>
-                    <Outlet />
-                </Content>
-            </Layout>
-        </Layout>
-    )
+.container {
+  display: flex;
+  flex-direction: row;
+  height: 100%;
 }
-export default LayoutRoot
+
+.divider {
+  position: fixed;
+  height: 100%;
+}
